# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Conflux Web Monitoring API
  description: REST API for monitoring and controlling Conflux orchestration
  contact:
    name: Conflux
    url: https://github.com/anomalyco/conflux
  license:
    name: MIT
    identifier: MIT
  version: 0.6.205
paths:
...
  /api/v2/events:
    get:
      tags:
      - remote-control
      summary: Server-Sent Events transport.
      description: |-
        Authenticated browser clients must consume this with `fetch()` response
        streaming; native `EventSource` cannot attach the bearer token and is not
        supported.
      operationId: events
      parameters:
      - name: after_sequence
        in: query
        description: Resume strictly after this event sequence
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: instance_id
        in: query
        description: Incarnation the cursor belongs to
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Ordered SSE event stream (fetch() response streaming)
        '401':
          description: Missing or invalid bearer credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/v2/ws:
    get:
      tags:
      - remote-control
      summary: WebSocket transport for non-browser clients.
      description: |-
        Authentication already happened in the v2 middleware, which also rejected any
        token supplied through the query string or `Sec-WebSocket-Protocol`, so by the
        time the upgrade runs the connection is known to be authorized.
      operationId: ws
      parameters:
      - name: after_sequence
        in: query
        description: Resume strictly after this event sequence
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: instance_id
        in: query
        description: Incarnation the cursor belongs to
        required: false
        schema:
          type: string
      responses:
        '101':
          description: Switching protocols
        '401':
          description: Missing bearer header, or credentials in query/subprotocol
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /ws:
    get:
      tags:
      - websocket
      summary: WebSocket connection handler for real-time state updates
      description: |-
        This endpoint upgrades the HTTP connection to a WebSocket connection
        and streams state updates to the client in real-time.

        # Protocol
        - Upon connection, the server sends an initial state message
        - Subsequently, the server broadcasts state updates as they occur
        - The client can send ping frames to keep the connection alive
        - The server responds with pong frames automatically

        # Message Format
        All messages are JSON-encoded with the following structure:
        ```json
        {
          "type": "initial_state" | "state_update",
          "timestamp": "ISO 8601 timestamp",
          "state": { ... }  // OrchestratorStateSnapshot
        }
        ```
      operationId: ws_handler
      responses:
        '101':
          description: Switching Protocols - WebSocket connection established
        '400':
          description: Bad Request - Not a valid WebSocket upgrade request
components:
  schemas:
    ApiError:
      type: object
      description: Structured error body returned by every non-success v2 response.
      required:
      - error_code
      - message
      - correlation_id
      properties:
        correlation_id:
          type: string
          description: Correlation ID of the failing request (caller-supplied or generated).
        current_revision:
          type:
          - integer
          - 'null'
          format: int64
          description: Current state revision, when the failure is revision-related.
          minimum: 0
        error_code:
          $ref: '#/components/schemas/ErrorCode'
          description: Stable code the client branches on.
        message:
          type: string
          description: Sanitized human-readable detail.
...
    ErrorCode:
      type: string
      description: |-
        Stable, machine-actionable v2 error codes.

        The HTTP status classifies the transport failure; the client branches on
        `error_code`. Two different causes may share status 409, so the code — not
        the status — is the contract.
      enum:
      - unauthorized
      - forbidden
      - not_found
      - stale_revision
      - lifecycle_conflict
      - target_ineligible
      - root_busy
      - idempotency_mismatch
      - registry_capacity
      - validation_failed
      - internal_error
...